requirements.txt
streamlit==1.41.1
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3